requirements.txt
# Essential production dependencies for Gradio Bailian Multi-Modal Video Application
gradio==5.43.1
requests>=2.28.0
python-dotenv>=1.0.0
dashscope>=1.14.0
Pillow>=10.0.0
oss2>=2.18.0